requirements.txt
numpy
pyyaml
tensorflow_datasets
protobuf
